+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Develop]
+
+### References
+
+- gnomad: r3.0 -> r3.1.1
+- [NEW] gnomad mt: r3.1
+
+## [10.0.3]
+
+- Remove duplicates from Glnexus output
+
+## [10.0.2]
+
+- Glnexus are used to genotype the gvcf regardless of how many samples that are analysed.
+- Resource bump to the MIP RNA recipe fusion_report
+
 ## [10.0.1]
 
 - Fix to gene panel regexp
@@ -10,8 +26,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Increased memory allocation for version_collect
 
 ### Tools
-stranger: 0.7.1 -> 0.8.0
 
+stranger: 0.7.1 -> 0.8.0
 
 ## [10.0.0]